REFACTOR ProductDeleteServiceImpl.java -> Handle delete process exceptions.

+++ src/main/java/org/example/simpleerp/service/product/impl/ProductDeleteServiceImpl.java
@@ -2,6 +2,7 @@ package org.example.simpleerp.service.product.impl;
 
 import lombok.RequiredArgsConstructor;
 import org.example.simpleerp.exception.product.ProductNotFoundException;
+import org.example.simpleerp.exception.product.UnableToDeleteProductException;
 import org.example.simpleerp.model.entity.ProductEntity;
 import org.example.simpleerp.repository.ProductRepository;
 import org.example.simpleerp.service.product.ProductDeleteService;
@@ -20,6 +21,14 @@ public class ProductDeleteServiceImpl implements ProductDeleteService {
         ProductEntity productEntityToBeDelete = productRepository
                 .findById(productId)
                 .orElseThrow(ProductNotFoundException::new);
+
+        if (Boolean.FALSE.equals(
+                productEntityToBeDelete.getOrderProductEntities().isEmpty())
+        ) {
+            throw new UnableToDeleteProductException(
+                    "Because its related with some Orders!"
+            );
+        }
 
         productRepository.delete(productEntityToBeDelete);
     }
